# Use AVX512 intrinsics for number conversion when possible