@martinothamar 2023-07-22
I guess this could be done in the JIT as well? Not sure what is preferred/intended https://github.com/dotnet/runtime/blob/08a6e06d8fdc29efebc5c0ec7d2b6ba215810e71/src/coreclr/jit/simdashwintrinsic.cpp#L533-L545