@adamsitnik 2023-10-20
> I'd prefer this to be in the JIT so that a mainline path (which compiles down to 1 instruction) doesn't end up pessimized by inlining heuristics

@tannergooding could you please provide some hints about how to achieve that? Like links to some sample code that does that for one simple instruction?