@martinothamar 2023-10-23
> I'd prefer this to be in the JIT so that a mainline path (which compiles down to 1 instruction) doesn't end up pessimized by inlining heuristics

Makes sense, I'll have a look some time this week 👍